the last touch
I added two transistors. The first one replaces the optocoupler and controls the relay: when 3.3 V is present at the input (which indicates a fault), the transistor switches on and activates the relay. If no voltage is present, the relay remains off.

The second transistor is used to turn the charger on and off. For this purpose, it was necessary to connect 12 V to its collector.